added functionality for delete button

+++ LogicUniversityTeam5/Controllers/RequisitionController.cs
@@ -16,6 +16,7 @@ namespace LogicUniversityTeam5.Controllers.Requisition
         IStockManagementService iStockService;
         IClassificationService iClassService;
         IRequisitionService iRequisitionService;
+        static int reqIdgen = 0;
         public RequisitionController(StockManagementService sms, ClassificationService cs,RequisitionService rs)
         {
             iStockService = sms;
@@ -31,14 +32,16 @@ namespace LogicUniversityTeam5.Controllers.Requisition
         public ActionResult ManageRequisition()
         {
             return View();
-        }
-
-      
+        }     
 
         public ActionResult ViewStationeryCatalogue()
         {
+            if (TempData["passedmodel"] != null)
+            {
+                TempData.Keep("passedmodel");
+            };
+
             CombinedViewModel combinedView = new CombinedViewModel();
-           
 
             combinedView.Items = iStockService.getAllItems();         
 
@@ -71,92 +74,50 @@ namespace LogicUniversityTeam5.Controllers.Requisition
                 List<string> reqDetailsItem = reqDetail.Select(r => r.ItemID).ToList();
                 List<Item> itemsAlreadyRequested = model.Items.Where(i => reqDetailsItem.Contains(i.ItemID)).ToList();
                 List<Item> newItemsForRequest = model.Items.Where(i => !reqDetailsItem.Contains(i.ItemID)).ToList();
-                    
-                for(int i=0;i<model.Items.Count;i++)
+                
+                for (int i = 0; i < model.Items.Count; i++)
                 {
                     string textBoxValue = model.AddedText[i].ToString().Trim();
                     if (reqDetailsItem.Contains(model.Items[i].ItemID) && (textBoxValue != null && textBoxValue != ""))
                     {
-                        //string textBoxValue = model.AddedText[i].ToString().Trim();
+
                         RequisitionDetail rd =
                             req.RequisitionDetails.First(x => x.ItemID.Equals(model.Items[i].ItemID));
                         rd.Quantity += Convert.ToInt32(textBoxValue);
                         passedmodel.Requisitions.Add(rd);
                     }
-                    else
-                    {
-                       //string textBoxValue = model.AddedText[i].ToString().Trim();
-                        if (textBoxValue != null && textBoxValue != "")
-                        {
-                            RequisitionDetail newRD = new RequisitionDetail();
-                            newRD.RequisitionID = req.RequisitionID;
-                            newRD.RequisitionDetailsID = -1; //Still not added to database.
-                            newRD.ItemID = model.Items[i].ItemID;
-                            newRD.Quantity = Convert.ToInt32(textBoxValue);
-                            newRD.Item = new Item();
-                            newRD.Item.ItemName = model.Items[i].ItemName;
-                            newRD.Item.UnitOfMeasure = model.Items[i].UnitOfMeasure;
 
-                            passedmodel.Requisitions.Add(newRD);
-                        }
-                        
+                    else if (textBoxValue != null && textBoxValue != "")
+                    {
+
+
+                        RequisitionDetail newRD = new RequisitionDetail();
+                        newRD.RequisitionID = req.RequisitionID;
+                        newRD.RequisitionDetailsID = --i; //Still not added to database.
+                        newRD.ItemID = model.Items[i].ItemID;
+                        newRD.Quantity = Convert.ToInt32(textBoxValue);
+                        newRD.Item = new Item();
+                        newRD.Item.ItemName = model.Items[i].ItemName;
+                        newRD.Item.UnitOfMeasure = model.Items[i].UnitOfMeasure;
+
+                        passedmodel.Requisitions.Add(newRD);
                     }
                 }
-                //foreach (RequisitionDetail rd in reqDetail)
-                //{
-                //    for(int i=0;i<model.AddedText.Count;i++)
-                //    {
-                //        string textBoxValue = model.AddedText[i].ToString().Trim();
-                //        if (!textBoxValue.Equals(null) || !textBoxValue.Equals(""))
-                //        {
-                //            if (model.Items[i].ItemID.Equals(rd.ItemID)){
-                //                rd.Quantity = rd.Quantity + Convert.ToInt32(textBoxValue);
-                //                passedmodel.Requisitions.Add(rd);
-                //            }
-                //            else if(!itemidList.Contains(model.Items[i].ItemID))
-                //            {
-                //                RequisitionDetail newRD = new RequisitionDetail();
-                //                newRD.RequisitionID = req.RequisitionID;
-                //                newRD.RequisitionDetailsID = - 1; //Still not added to database.
-                //                newRD.ItemID = model.Items[i].ItemID;
-                //                newRD.Quantity = Convert.ToInt32(model.AddedText[i]);
-                //                newRD.Item = new Item();
-                //                newRD.Item.ItemName = model.Items[i].ItemName;
-                //                newRD.Item.UnitOfMeasure = model.Items[i].UnitOfMeasure;
-                //                itemidList.Add(newRD.ItemID);
-                //                newRDlist.Add(newRD);
-                //            }
-                //        }
-
-                //     }
-                    
-
-                // }
-                }
-
-            //for (int i = 0; i < model.AddedText.Count; i++)
-            //{
-            //    string textBoxValue = model.AddedText[i].ToString();
-            //    string trimmed = textBoxValue.Trim();
-
-            //    if (!(trimmed.Equals(null)) && !trimmed.Equals(""))
-            //    {                    
-            //        Item item = new Item()
-            //        {
-            //            ItemID = model.Items[i].ItemID,
-            //            ItemName = model.Items[i].ItemName,
-            //            UnitOfMeasure = model.Items[i].UnitOfMeasure
-            //        };
-
-            //passedmodel.Items.Add(item);
-            //passedmodel.AddedText.Add(trimmed);
-
-            //    }
-
-
-            //}
-           
-
+                foreach(Item i in itemsAlreadyRequested)
+                {
+                    if(!newItemsForRequest.Contains(i))
+                    {
+                        RequisitionDetail oldReq= req.RequisitionDetails.First(x => x.ItemID.Equals(i.ItemID));
+                        passedmodel.Requisitions.Add(oldReq);
+                    }
+                }                 
+                
+                     
+                
+             }         
+            
+                                 
+               
             TempData["passedmodel"] = passedmodel;
             TempData.Keep("passedmodel");
             return RedirectToAction("StationeryRequestForm");
@@ -167,18 +128,6 @@ namespace LogicUniversityTeam5.Controllers.Requisition
         {
             CombinedViewModel newmodel = (CombinedViewModel)TempData["passedmodel"];
             TempData.Keep("passedmodel");
-
-            //ServiceLayer.DataAccess.Requisition req = iRequisitionService.getUnsubmittedRequisitionOfEmployee("E001");
-            //List<RequisitionDetail> reqDetail ;
-            
-            //if (req != null)
-            //{
-            //    reqDetail = iRequisitionService.getRequisitionDetails(req.RequisitionID);
-            //    if (reqDetail != null)
-            //    {
-            //            newmodel.Requisitions=reqDetail;      
-            //    }      
-            //}
            
             newmodel.IsSave = false;
             return View(newmodel);
@@ -199,7 +148,7 @@ namespace LogicUniversityTeam5.Controllers.Requisition
                     {
                         iRequisitionService.addNewRequisitionDetail(req.RequisitionID, model.Requisitions[i].ItemID, Convert.ToInt32(textBoxValue));
                     }
-                    else
+                    else if (model.Requisitions[i].RequisitionDetailsID >= 0)
                     {
                         iRequisitionService.editRequisitionDetailQty(model.Requisitions[i].RequisitionDetailsID, Convert.ToInt32(textBoxValue));
                     }
@@ -215,7 +164,7 @@ namespace LogicUniversityTeam5.Controllers.Requisition
                 iRequisitionService.submitRequisition(req.RequisitionID);
                 return RedirectToAction("Index", "Home");
             }
-         //   iRequisitionService.submitRequisition(req.RequisitionID);
+       
             
         }
 
@@ -223,35 +172,67 @@ namespace LogicUniversityTeam5.Controllers.Requisition
         [HttpGet]
         public ActionResult ResubmitStationaryRequestForm()
         {
-            CombinedViewModel combinedView = new CombinedViewModel();
+            CombinedViewModel combinedView = new CombinedViewModel();            
             combinedView.AddedText = new List<string>();
-            ServiceLayer.DataAccess.Requisition  req =   iRequisitionService.getUnsubmittedRequisitionOfEmployee("E001");        
-            int count = req.RequisitionID;
+            ServiceLayer.DataAccess.Requisition  req =   iRequisitionService.getUnsubmittedRequisitionOfEmployee("E001");      
+            
             List<RequisitionDetail> reqdetails = iRequisitionService.getRequisitionDetails(req.RequisitionID);
             combinedView.Requisitions = reqdetails;
             for(int i=0;i<reqdetails.Count;i++)
             {
                 combinedView.AddedText.Add(Convert.ToString(reqdetails[i].Quantity));
             }
-
+            TempData["passmodel"] = combinedView;
             return View(combinedView);
         }
 
         [HttpGet]
         public ActionResult Delete(int id)
         {
-            iRequisitionService.deleteRequisitionDetail(id);
-            //CombinedViewModel combinedView = new CombinedViewModel();
+
+
             CombinedViewModel combinedView = (CombinedViewModel)TempData["passedmodel"];
-          //  combinedView.AddedText = new List<string>();
+            if (combinedView != null)
+            {
+                List<RequisitionDetail> rDetailList = combinedView.Requisitions;
+
+
+
+                foreach (RequisitionDetail rd in rDetailList.ToList())
+                {
+                    
+                        rDetailList.Remove(rd);
+                }
+
+                combinedView.Requisitions = rDetailList;
+                //foreach (RequisitionDetail rd in rDetailList.ToList())
+                //{
+                //    combinedView.Requisitions.Add(rd);
+                //}
+            }
+            else
+            {
+                combinedView = new CombinedViewModel();
+                combinedView.AddedText = new List<string>();
+                combinedView.Requisitions = new List<RequisitionDetail>();
+
+            }
+
+            //Delete from existing reqDetails in DB
+            if (id > 0)
+            {
+                iRequisitionService.deleteRequisitionDetail(id);
+            }
+
+            //combinedView.AddedText = new List<string>();
+            //combinedView.Requisitions = new List<RequisitionDetail>();
             ServiceLayer.DataAccess.Requisition req = iRequisitionService.getUnsubmittedRequisitionOfEmployee("E001");
-           // int count = req.RequisitionID;
             List<RequisitionDetail> reqdetails = iRequisitionService.getRequisitionDetails(req.RequisitionID);
-            combinedView.Requisitions = reqdetails;
+
             for (int i = 0; i < reqdetails.Count; i++)
             {
-                
-                combinedView.AddedText.Add(Convert.ToString(reqdetails[i].Quantity));
+                //combinedView.AddedText.Add(Convert.ToString(reqdetails[i].Quantity));
+                combinedView.Requisitions.Add(reqdetails[i]);
             }
             TempData["passedmodel"] = combinedView;
             return RedirectToAction("StationeryRequestForm");
